% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletrs-package.R
\docType{package}
\name{doubletrs-package}
\alias{doubletrs}
\alias{doubletrs-package}
\title{doubletrs: Memory-Efficient Doublet Detection for Single-Cell RNA Sequencing}
\description{
Detects doublets in single-cell RNA sequencing data with the 'DoubletFinder' method of McGinnis, Murrow and Gartner (2019) \doi{10.1016/j.cels.2019.03.003}, reimplemented in 'Rust' to use less memory and time. Provides artificial doublet simulation, exact nearest-neighbour search and pANN (proportion of artificial nearest neighbours) scoring, and reproduces 'DoubletFinder' scores exactly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mayankgandhi13/doublet-rs}
  \item Report bugs at \url{https://github.com/mayankgandhi13/doublet-rs/issues}
}

}
\author{
\strong{Maintainer}: Mayank Gandhi \email{gandhi.may@northeastern.edu} [copyright holder]

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS for the bundled crates, their authors and licenses) [copyright holder]
}

}
\keyword{internal}
